+++ eras/allegra/impl/src/Cardano/Ledger/Allegra/Translation.hs
@@ -156,7 +156,7 @@ instance TranslateEra AllegraEra EpochState where
   translateEra ctxt es =
     return
       EpochState
-        { esAccountState = esAccountState es
+        { esChainAccountState = esChainAccountState es
         , esSnapshots = esSnapshots es
         , esLState = translateEra' ctxt $ esLState es
         , esNonMyopic = esNonMyopic es