+++ eras/conway/impl/src/Cardano/Ledger/Conway/Rules/Epoch.hs
@@ -67,10 +67,7 @@ import Cardano.Ledger.Shelley.LedgerState (
   EpochState (..),
   LedgerState (..),
   UTxOState (..),
-  asTreasuryL,
   curPParamsEpochStateL,
-  esAccountState,
-  esAccountStateL,
   esLStateL,
   esSnapshotsL,
   lsCertStateL,
@@ -228,23 +225,23 @@ updateNumDormantEpochs currentEpoch ps vState =
 --
 -- The utxo fees and donations are applied in the remaining body of EPOCH transition
 applyEnactedWithdrawals ::
-  AccountState ->
+  ChainAccountState ->
   DState era ->
   EnactState era ->
-  (AccountState, DState era, EnactState era)
-applyEnactedWithdrawals accountState dState enactedState =
+  (ChainAccountState, DState era, EnactState era)
+applyEnactedWithdrawals chainAccountState dState enactedState =
   let enactedWithdrawals = enactedState ^. ensWithdrawalsL
       rewardsUView = RewDepUView $ dState ^. dsUnifiedL
       successfulWithdrawls = rewardsUView ◁ enactedWithdrawals
-      accountState' =
-        accountState
+      chainAccountState' =
+        chainAccountState
           -- Subtract `successfulWithdrawals` from the treasury, and add them to the rewards UMap
           -- `unclaimed` withdrawals remain in the treasury.
           -- Compared to the spec, instead of adding `unclaimed` and subtracting `totWithdrawals`
           --   + unclaimed - totWithdrawals
           -- we just subtract the `refunds`
           --   - refunds
-          & asTreasuryL %~ (<-> fold successfulWithdrawls)
+          & casTreasuryL %~ (<-> fold successfulWithdrawls)
       -- The use of the partial function `compactCoinOrError` is justified here because
       -- 1. the decoder for coin at the proposal-submission boundary has already
       --    confirmed we have a compactible value
@@ -258,7 +255,7 @@ applyEnactedWithdrawals accountState dState enactedState =
         enactedState
           & ensWithdrawalsL .~ Map.empty
           & ensTreasuryL .~ mempty
-   in (accountState', dState', enactedState')
+   in (chainAccountState', dState', enactedState')
 
 epochTransition ::
   forall era.
@@ -291,14 +288,14 @@ epochTransition = do
   TRC
     ( ()
       , epochState0@EpochState
-          { esAccountState = accountState0
-          , esSnapshots = snapshots0
+          { esSnapshots = snapshots0
           , esLState = ledgerState0
           }
       , eNo
       ) <-
     judgmentContext
-  let govState0 = utxosGovState utxoState0
+  let chainAccountState0 = epochState0 ^. chainAccountStateL
+      govState0 = utxosGovState utxoState0
       curPParams = govState0 ^. curPParamsGovStateL
       utxoState0 = lsUTxOState ledgerState0
       certState0 = ledgerState0 ^. lsCertStateL
@@ -314,9 +311,9 @@ epochTransition = do
           { psStakePoolParams = newStakePoolParams
           , psFutureStakePoolParams = Map.empty
           }
-  PoolreapState utxoState1 accountState1 certState1 <-
+  PoolreapState utxoState1 chainAccountState1 certState1 <-
     trans @(EraRule "POOLREAP" era) $
-      TRC ((), PoolreapState utxoState0 accountState0 (certState0 & certPStateL .~ pState1), eNo)
+      TRC ((), PoolreapState utxoState0 chainAccountState0 (certState0 & certPStateL .~ pState1), eNo)
 
   let
     stakePoolDistr = ssStakeMarkPoolDistr snapshots1
@@ -325,8 +322,8 @@ epochTransition = do
     ratifyState@RatifyState {rsEnactState, rsEnacted, rsExpired} =
       extractDRepPulsingState pulsingState
 
-    (accountState2, dState2, EnactState {..}) =
-      applyEnactedWithdrawals accountState1 (certState1 ^. certDStateL) rsEnactState
+    (chainAccountState2, dState2, EnactState {..}) =
+      applyEnactedWithdrawals chainAccountState1 (certState1 ^. certDStateL) rsEnactState
 
     -- NOTE: It is important that we apply the results of ratification
     -- and enactment from the pulser to the working copy of proposals.
@@ -375,10 +372,10 @@ epochTransition = do
         )
         (certState1 ^. certPStateL)
         (dState2 & dsUnifiedL .~ newUMap)
-    accountState3 =
-      accountState2
+    chainAccountState3 =
+      chainAccountState2
         -- Move donations and unclaimed rewards from proposals to treasury:
-        & asTreasuryL <>~ (utxoState0 ^. utxosDonationL <> fold unclaimed)
+        & casTreasuryL <>~ (utxoState0 ^. utxosDonationL <> fold unclaimed)
     utxoState2 =
       utxoState1
         & utxosDepositedL .~ totalObligation certState2 govState1
@@ -391,7 +388,7 @@ epochTransition = do
         & lsUTxOStateL .~ utxoState2
     epochState1 =
       epochState0
-        & esAccountStateL .~ accountState3
+        & chainAccountStateL .~ chainAccountState3
         & esSnapshotsL .~ snapshots1
         & esLStateL .~ ledgerState1
   tellEvent $ EpochBoundaryRatifyState ratifyState