+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/API/Validation.hs
@@ -213,7 +213,7 @@ mkBbodyEnv
     } =
     STS.BbodyEnv
       { STS.bbodyPp = nesEs ^. curPParamsEpochStateL
-      , STS.bbodyAccount = LedgerState.esAccountState nesEs
+      , STS.bbodyAccount = LedgerState.esChainAccountState nesEs
       }
 
 updateNewEpochState ::