+++ libs/cardano-ledger-conformance/src/Test/Cardano/Ledger/Conformance/SpecTranslate/Conway/Ledger.hs
@@ -15,8 +15,8 @@ module Test.Cardano.Ledger.Conformance.SpecTranslate.Conway.Ledger () where
 import Cardano.Ledger.BaseTypes (Inject)
 import Cardano.Ledger.Conway.Core (EraPParams (..), EraRule, ScriptHash)
 import Cardano.Ledger.Conway.Rules (ConwayLedgerPredFailure, EnactState)
-import Cardano.Ledger.Shelley.LedgerState (AccountState (..))
 import Cardano.Ledger.Shelley.Rules (LedgerEnv (..))
+import Cardano.Ledger.Shelley.State (ChainAccountState (..))
 import Control.State.Transition.Extended (STS (..))
 import Data.Functor.Identity (Identity)
 import Data.Maybe.Strict (StrictMaybe)
@@ -44,7 +44,7 @@ instance
       <*> toSpecRep policyHash
       <*> toSpecRep ledgerPp
       <*> toSpecRep enactState
-      <*> toSpecRep (asTreasury ledgerAccount)
+      <*> toSpecRep (casTreasury ledgerAccount)
 
 instance
   ( ToExpr (PredicateFailure (EraRule "GOV" era))