+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Constrained/Conway/LedgerTypes/WellFormed.hs
@@ -1,7 +1,6 @@
 {-# LANGUAGE AllowAmbiguousTypes #-}
 {-# LANGUAGE CPP #-}
 {-# LANGUAGE DataKinds #-}
-{-# LANGUAGE ExplicitForAll #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
@@ -10,7 +9,6 @@
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE TypeOperators #-}
-{-# LANGUAGE TypeSynonymInstances #-}
 {-# LANGUAGE UndecidableInstances #-}
 {-# LANGUAGE UndecidableSuperClasses #-}
 
@@ -62,8 +60,8 @@ import Test.QuickCheck (Gen)
 ppX :: forall era. EraSpecPParams era => Gen (PParams era)
 ppX = genFromSpec @ConwayFn @(PParams era) pparamsSpec
 
-acctX :: Gen AccountState
-acctX = genFromSpec @ConwayFn @AccountState accountStateSpec
+acctX :: Gen ChainAccountState
+acctX = genFromSpec @ConwayFn @ChainAccountState accountStateSpec
 
 psX :: forall era. GenScript era => Gen (PState era)
 psX = do
@@ -74,7 +72,7 @@ psX = do
 dsX :: forall era. EraSpecLedger era ConwayFn => Gen (DState era)
 dsX = do
   univ <- genWitUniv 25
-  acct <- genFromSpec @ConwayFn @AccountState accountStateSpec
+  acct <- genFromSpec @ConwayFn @ChainAccountState accountStateSpec
   pools <-
     genFromSpec @ConwayFn @(Map (KeyHash 'StakePool) PoolParams)
       (hasSize (rangeSize 8 8))
@@ -94,7 +92,7 @@ vsX = do
 csX :: forall era. EraSpecLedger era ConwayFn => Gen (CertState era)
 csX = do
   univ <- genWitUniv 25
-  acct <- genFromSpec @ConwayFn @AccountState accountStateSpec
+  acct <- genFromSpec @ConwayFn @ChainAccountState accountStateSpec
   epoch <- genFromSpec @ConwayFn @EpochNo epochNoSpec
   genFromSpec @ConwayFn @(CertState era)
     (certStateSpec univ (lit acct) (lit epoch))
@@ -136,7 +134,7 @@ lsX ::
   PParams era -> Gen (LedgerState era)
 lsX pp = do
   univ <- genWitUniv @era 50
-  acct <- genFromSpec @ConwayFn @AccountState accountStateSpec
+  acct <- genFromSpec @ConwayFn @ChainAccountState accountStateSpec
   epoch <- genFromSpec @ConwayFn @EpochNo epochNoSpec
   genFromSpec @ConwayFn @(LedgerState era) (ledgerStateSpec pp univ (lit acct) (lit epoch))
 
@@ -170,7 +168,7 @@ snapsX ::
   PParams era -> Gen SnapShots
 snapsX pp = do
   univ <- genWitUniv @era 50
-  acct <- genFromSpec @ConwayFn @AccountState accountStateSpec
+  acct <- genFromSpec @ConwayFn @ChainAccountState accountStateSpec
   epoch <- genFromSpec @ConwayFn @EpochNo epochNoSpec
   ls <- genFromSpec @ConwayFn @(LedgerState era) (ledgerStateSpec pp univ (lit acct) (lit epoch))
   genFromSpec @ConwayFn @SnapShots (snapShotsSpec (lit (getMarkSnapShot ls)))
@@ -178,7 +176,7 @@ snapsX pp = do
 instanRewX :: forall era. EraSpecTxOut era ConwayFn => Gen InstantaneousRewards
 instanRewX = do
   univ <- genWitUniv @era 50
-  acct <- genFromSpec @ConwayFn @AccountState accountStateSpec
+  acct <- genFromSpec @ConwayFn @ChainAccountState accountStateSpec
   genFromSpec @ConwayFn @InstantaneousRewards
     (irewardSpec @era @ConwayFn univ (lit acct))
 
@@ -210,7 +208,7 @@ instance EraSpecPParams era => WellFormed (PParams era) era where
   wff = ppX @era
   wffWithPP = pure
 
-instance EraSpecPParams era => WellFormed AccountState era where
+instance EraSpecPParams era => WellFormed ChainAccountState era where
   wff = acctX
   wffWithPP _ = acctX